requirements.txt
Babel==2.4.0
blinker==1.4
click==6.7
decorator==4.0.11
defusedxml==0.5.0
Flask==0.12.2
Flask-HTTPAuth==3.2.3
guess-language==0.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pbr==3.0.1
PyJWT==1.5.0
python3-openid==3.1.0
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
